rename workspace.js to Project
